+++ src/components/CardProfile.css
@@ -9,59 +9,60 @@
 	background-color: var(--background-color-header-footer);
 	padding: 10px 0px;
 	border-radius: 7px;
+}
 
-	.cardTop {
-		display: flex;
-		justify-content: space-between;
-		align-items: center;
-		width: 100%;
-		padding: 0px 10px;
+.cardTop {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	width: 100%;
+	padding: 0px 10px;
+}
 
-		.infoDog {
-			display: flex;
-			flex-direction: column;
-			align-items: flex-start;
-			color: var(--main-color);
-		}
+.infoDog {
+	display: flex;
+	flex-direction: column;
+	align-items: flex-start;
+	color: var(--main-color);
+}
 
-		.raceCityDog {
-			font-size: 15px;
-			--text-color: black;
-			color: var(--main-color);
-		}
-	}
+.raceCityDog {
+	font-size: 15px;
+	color: var(--main-color);
+}
 
-	.descriptionsDog {
-		padding: 0px 10px;
+.descriptionsDog {
+	padding: 0px 10px;
+}
 
-		.descriptionDog {
-			display: flex;
-			padding-bottom: 7px;
+.descriptionDog {
+	display: flex;
+	padding-bottom: 7px;
+}
 
-			p {
-				text-align: left;
-				line-height: 1, 5;
-				font-size: 12px;
-			}
+.descriptionDog p {
+	text-align: left;
+	line-height: 1.5;
+	font-size: 12px;
+}
 
-			.iconeDescriptionDog {
-				font-size: 20px;
-				padding-right: 5px;
-			}
-		}
-	}
+.iconeDescriptionDog {
+	font-size: 20px;
+	padding-right: 5px;
+}
 
-	.photosDogs {
-		width: 100%;
-		height: 50%;
-	}
+.photosDogs {
+	width: 100%;
+	height: 50%;
+	object-fit: cover;
+	border-radius: 5px;
+}
 
-	.likeDislikeButtons {
-		display: flex;
-		justify-content: space-evenly;
-		width: 100%;
-		padding-bottom: 20px;
-	}
+.likeDislikeButtons {
+	display: flex;
+	justify-content: space-evenly;
+	width: 100%;
+	padding-bottom: 20px;
 }
 
 button {
