+++ src/components/CardProfileTrash.css
@@ -1,4 +1,5 @@
-.cardProfile {
+.cardProfile,
+.cardProfileback {
 	display: flex;
 	flex-direction: column;
 	align-items: center;
@@ -7,83 +8,13 @@
 	height: 450px;
 	background-color: #efede6;
 	color: black;
-	padding-bottom: 10px;
-	gap: 20px;
-	border-radius: 5px;
-	box-shadow: 5px 5px 0px grey;
-
-	.cardTop {
-		display: flex;
-		justify-content: space-between;
-		align-items: center;
-		width: 100%;
-
-		#infoDog {
-			display: flex;
-			flex-direction: column;
-			align-items: flex-start;
-			color: #5c913b;
-
-			#nameDog {
-				font-weight: bold;
-			}
-		}
-	}
-
-	.logosProfile {
-		display: flex;
-		justify-content: space-evenly;
-		width: 100%;
-	}
-
-	#photoDog {
-		border-radius: 5px;
-	}
-}
-
-.cardProfileback {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: space-between;
-	width: 280px;
-	height: 485px;
-	background-color: #efede6;
-	color: black;
 	padding: 10px;
 	gap: 20px;
 	border-radius: 5px;
 	box-shadow: 5px 5px 0px grey;
-
-	.cardTop {
-		display: flex;
-		justify-content: space-between;
-		align-items: center;
-		width: 100%;
-
-		#infoDog {
-			display: flex;
-			flex-direction: column;
-			align-items: flex-start;
-			color: #5c913b;
-
-			#nameDog {
-				font-weight: bold;
-			}
-		}
-	}
-
-	.logosProfile {
-		display: flex;
-		justify-content: space-evenly;
-		width: 100%;
-	}
-
-	#photoDog {
-		border-radius: 5px;
-	}
 }
 
+/* Animations sur les cartes lors du retrait */
 .cardProfile {
 	transition: opacity 0.4s ease, transform 0.4s ease;
 	opacity: 1;
@@ -98,7 +29,89 @@
 	pointer-events: none;
 }
 
-/* Cache la face arrière pendant la suppression */
 .cardProfile.removing:nth-child(2) {
 	display: none;
+}
+
+/* Haut de la carte */
+.cardTop {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	width: 100%;
+}
+
+/* Informations du chien */
+.infoDog {
+	display: flex;
+	flex-direction: column;
+	align-items: flex-start;
+	color: #5c913b;
+}
+
+.nameDog {
+	font-weight: bold;
+}
+
+/* Logos (like / dislike) */
+.logosProfile {
+	display: flex;
+	justify-content: space-evenly;
+	width: 100%;
+}
+
+/* Photo du chien */
+.photoDog {
+	border-radius: 5px;
+	object-fit: cover;
+	width: 100%;
+	height: auto;
+	max-height: 250px;
+}
+
+/* Description */
+.descriptionsDog {
+	width: 100%;
+}
+
+.descriptionDog {
+	display: flex;
+	align-items: flex-start;
+	margin-bottom: 8px;
+	padding: 0 5px;
+}
+
+.iconeDescriptionDog {
+	font-size: 20px;
+	margin-right: 5px;
+}
+
+.descriptionDog p {
+	text-align: left;
+	line-height: 1.5;
+	font-size: 12px;
+	margin: 0;
+}
+
+/* Boutons like / dislike */
+.likeDislikeButtons {
+	display: flex;
+	justify-content: space-evenly;
+	width: 100%;
+	padding-bottom: 10px;
+}
+
+button {
+	border: none;
+	background: none;
+	padding: 0;
+	margin: 0;
+	box-shadow: none;
+	outline: none;
+	cursor: pointer;
+}
+
+.iconFlipCard {
+	width: 30px;
+	height: 30px;
 }
