+++ src/components/LikedDogsList/LikedDogsList.css
@@ -1,0 +1,6 @@
+.noLikesMessage {
+	text-align: center;
+	font-style: italic;
+	color: #888;
+	margin-top: 2rem;
+}
