--- src/components/Responsive.css
@@ -1,46 +1,0 @@
-/* Container des cartes */
-.cardList {
-	display: flex;
-	flex-wrap: wrap;
-	gap: 20px;
-	justify-content: center;
-	padding: 20px;
-}
-
-/* Chaque carte */
-.cardProfile {
-	width: 300px;
-	background-color: white;
-	border-radius: 12px;
-	box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
-	padding: 16px;
-	box-sizing: border-box;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-}
-
-/* Logos ou boutons */
-.logosProfile {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	gap: 16px;
-	margin-top: 16px;
-}
-
-/* Responsive : en dessous de 768px, une seule carte par ligne */
-@media screen and (max-width: 768px) {
-	.cardList {
-		flex-direction: column;
-		align-items: center;
-	}
-
-	.cardProfile {
-		width: 90%; /* occupe 90% de la largeur de l'écran */
-	}
-
-	.logosProfile {
-		justify-content: center;
-	}
-}
