+++ src/pages/home/Home.tsx
@@ -1,7 +1,5 @@
 import "./Home.css";
 
-import Likes from "../../components/Likes";
-
 function Home() {
 	return <h1>Bienvenue sur Doggo</h1>;
 }
