--- src/reset.css
@@ -1,38 +1,0 @@
-*,
-*::before,
-*::after {
-	box-sizing: border-box;
-	margin: 0;
-	padding: 0;
-}
-
-body {
-	text-rendering: optimizeSpeed;
-	font-family: system-ui, sans-serif;
-}
-
-img,
-picture,
-video,
-canvas,
-svg {
-	display: block;
-	max-width: 100%;
-}
-
-input,
-button,
-textarea,
-select {
-	font: inherit;
-}
-
-a {
-	text-decoration: none;
-	color: inherit;
-}
-
-ul,
-ol {
-	list-style: none;
-}
